Move modules into lib, es6ify gulp files
