+++ .eslintrc
@@ -11,6 +11,7 @@
     "comma-dangle": ["error", "always-multiline"],
     "default-case": 0,
     "global-require": 0,
-    "import/no-unresolved": 0
+    "import/no-unresolved": 0,
+    "import/no-extraneous-dependencies": 0
   }
 }
