+++ src/api/src/lib/rules/min/index.js
