+++ src/api/src/lib/rules/password/index.js
