+++ src/api/src/lib/rules/password/index.spec.js
