+++ src/api/src/lib/rules/unique-email/index.spec.js
