+++ src/api/src/lib/rules/valid-gw2-token/index.spec.js
